Add BCD-Extension functionality: C-MassOps
Delete command: upgrade functionality

Delete command did not support deleting multiple entries in one
command

It is more user friendly to be capable of deleting multiple
entries in one go, as deleting changes the order of the list, which
may confuse the user.

Let's:

Update the delete command to take in multiple integer type
parameters.

Make it support the deletion of multiple list entries in one
command